requirements.txt
Flask
Werkzeug
tensorflow
numpy
Pillow
joblib 